+++ libnavui-speedlimit/src/main/java/com/mapbox/navigation/ui/speedlimit/ComponentInstaller.kt
@@ -1,10 +1,15 @@
 package com.mapbox.navigation.ui.speedlimit
 
+import android.content.Context
 import androidx.annotation.StyleRes
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
+import com.mapbox.navigation.base.formatter.DistanceFormatterOptions
 import com.mapbox.navigation.ui.base.installer.ComponentInstaller
 import com.mapbox.navigation.ui.base.installer.Installation
+import com.mapbox.navigation.ui.speedlimit.internal.SpeedInfoComponent
 import com.mapbox.navigation.ui.speedlimit.internal.SpeedLimitComponent
+import com.mapbox.navigation.ui.speedlimit.model.MapboxSpeedInfoOptions
+import com.mapbox.navigation.ui.speedlimit.view.MapboxSpeedInfoView
 import com.mapbox.navigation.ui.speedlimit.view.MapboxSpeedLimitView
 
 /**
@@ -29,6 +34,27 @@ fun ComponentInstaller.speedLimit(
 }
 
 /**
+ * Install components that render [MapboxSpeedInfoView].
+ *
+ * @param speedInfoView [MapboxSpeedInfoView]
+ * @param config SpeedInfoConfig
+ */
+@ExperimentalPreviewMapboxNavigationAPI
+fun ComponentInstaller.speedInfo(
+    speedInfoView: MapboxSpeedInfoView,
+    config: SpeedInfoConfig.() -> Unit = {}
+): Installation {
+    val componentConfig = SpeedInfoConfig(speedInfoView.context).apply(config)
+    return component(
+        SpeedInfoComponent(
+            speedInfoOptions = componentConfig.speedInfoOptions,
+            speedInfoView = speedInfoView,
+            distanceFormatterOptions = componentConfig.distanceFormatterOptions
+        )
+    )
+}
+
+/**
  * Speed limit view component configuration class.
  */
 @ExperimentalPreviewMapboxNavigationAPI
@@ -42,4 +68,21 @@ class SpeedLimitConfig internal constructor() {
      * [textAppearance] to be set to [MapboxSpeedLimitView].
      */
     @StyleRes var textAppearance: Int = 0
+}
+
+/**
+ * Speed info view component configuration class.
+ */
+@ExperimentalPreviewMapboxNavigationAPI
+class SpeedInfoConfig internal constructor(context: Context) {
+    /**
+     * [speedInfoOptions] to be set to [MapboxSpeedInfoView].
+     */
+    var speedInfoOptions: MapboxSpeedInfoOptions = MapboxSpeedInfoOptions.Builder().build()
+
+    /**
+     * [DistanceFormatterOptions] to be used by [MapboxSpeedInfoView].
+     */
+    var distanceFormatterOptions: DistanceFormatterOptions =
+        DistanceFormatterOptions.Builder(context).build()
 }